+++ TagCloudContainerTask/TagCloud/TagCloud.csproj
@@ -43,10 +43,10 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="WordsFiltering\WordsFilter.cs" />
-    <Compile Include="WordsPreprocessing\WordPreprocessor.cs" />
-    <Compile Include="WordsReading\InitialWordFormReader.cs" />
-    <Compile Include="WordsReading\IReader.cs" />
+    <Compile Include="Words\Filtering\WordsFilter.cs" />
+    <Compile Include="Words\Preprocessing\WordPreprocessor.cs" />
+    <Compile Include="Words\Reading\InitialWordFormReader.cs" />
+    <Compile Include="Words\Reading\IReader.cs" />
   </ItemGroup>
   <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
